requirements.txt
gcsfs
google-cloud-bigquery
unittest
pyyaml
